moved both the Input and Select components inside the FormComponent as they share styles.
